+++ openquake/qa_tests_data/event_based_damage/case_12/expected/d_damages.txt
