+++ openquake/qa_tests_data/event_based_damage/case_13/expected/c_damage_table.txt
