+++ openquake/qa_tests_data/event_based_damage/case_13/exposure1.xml
