+++ openquake/qa_tests_data/event_based_damage/case_14/consequences_by_taxonomy.csv
