+++ openquake/qa_tests_data/event_based_damage/case_14/expected/dmgcsq.csv
