+++ openquake/qa_tests_data/event_based_damage/case_14/structural_fragility_model.xml
